LayoutTests/fast/forms/date/date-editable-components/date-editable-components-input-and-change-events.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../resources/js-test.js"></script>
<script src="../../../../resources/ui-helper.js"></script>
</head>
<body>

<input id="input" type="date">

<script>

jsTestIsAsync = true;

changeEventsFired = 0;
function onChangeEvent() {
    changeEventsFired++;
}

inputEventsFired = 0;
function onInputEvent() {
    inputEventsFired++;
}

input.addEventListener("change", onChangeEvent);
input.addEventListener("input", onInputEvent);

addEventListener("load", async () => {
    description("Test for input and change events for &lt;input type=date&gt;");

    input.focus();                                         // [mm]/dd/yyyy
    UIHelper.keyDown("9");                                 // -> [09]/dd/yyyy
    shouldBe("inputEventsFired", "0");
    shouldBe("changeEventsFired", "0");

    UIHelper.keyDown("rightArrow");                        // -> 09/[dd]/yyyy
    UIHelper.keyDown("2");                                 // -> 09/[02]/yyyy
    shouldBe("inputEventsFired", "0");
    shouldBe("changeEventsFired", "0");

    UIHelper.keyDown("rightArrow");                        // -> 09/02/[yyyy]
    UIHelper.keyDown("2");                                 // -> 09/02/[0002]
    shouldBe("inputEventsFired", "1");
    shouldBe("changeEventsFired", "0");

    UIHelper.keyDown("0");                                 // -> 09/02/[0020]
    shouldBe("inputEventsFired", "2");
    shouldBe("changeEventsFired", "0");

    UIHelper.keyDown("1");                                 // -> 09/02/[0201]
    shouldBe("inputEventsFired", "3");
    shouldBe("changeEventsFired", "0");

    UIHelper.keyDown("2");                                 // -> 09/02/[2012]
    shouldBe("inputEventsFired", "4");
    shouldBe("changeEventsFired", "0");

    UIHelper.keyDown("A");                                 // Ignored.
    shouldBe("inputEventsFired", "4");
    shouldBe("changeEventsFired", "0");

    shouldBeEqualToString("input.value", "2012-09-02");

    input.blur();
    shouldBe("inputEventsFired", "4");
    shouldBe("changeEventsFired", "1");

    finishJSTest();
});

</script>
</body>
</html>
